server/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0a1
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
cssselect==1.1.0
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
django-rest-auth==0.9.5
django-sendgrid-v5==0.9.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
fake-useragent==0.1.11
feedparser==6.0.2
future==0.18.2
html5lib==1.0.1
idna==2.10
importlib-metadata==3.7.2
ipaddr==2.2.0
isort==5.7.0
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
lockfile==0.12.2
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.1
mccabe==0.6.1
msgpack==0.6.2
multitasking==0.0.9
mypy-extensions==0.4.3
numpy==1.20.2
oauthlib==3.1.0
packaging==20.3
pandas==1.2.3
parse==1.19.0
pathspec==0.8.1
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
pyee==8.1.0
PyJWT==2.0.1
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.6
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-http-client==3.3.2
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.24.1
scipy==1.6.2
sendgrid==6.6.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
w3lib==1.22.0
webencodings==0.5.1
websockets==8.1
wrapt==1.12.1
yahoo-fin==0.8.8
yfinance==0.1.55
zipp==3.4.1
torch==1.8.1
